refactor: extract partner dialog
