+++ src/professional_invoice_manager/pages.py
@@ -1,0 +1,1 @@
+"""Page widgets for management features."""
